requirements.txt
streamlit
requests
flask
google-generativeai
python-dotenv
sentence-transformers
torch